images and icons moved to assets
