+++ css/currency.css
@@ -157,7 +157,7 @@ img[data-src] {
   border-radius: 1.875rem;
   color: var(--white);
   background-color: var(--accent-blue);
-  background-image: url("../../images/svg/wave-mobile.svg");
+  background-image: url("../../assets/icons/wave-mobile.svg");
   background-size: contain;
   background-repeat: no-repeat;
   background-position: bottom;
@@ -166,13 +166,13 @@ img[data-src] {
   .currency {
     width: 100%;
     background-size: contain;
-    background-image: url("../../images/svg/wave-tablet.svg");
+    background-image: url("../../assets/icons/wave-tablet.svg");
   }
 }
 @media (min-width: 768px) and (max-width: 1279px) {
   .currency {
     max-width: 336px;
-    background-image: url("../../images/svg/wave-desktop.svg");
+    background-image: url("../../assets/icons/wave-desktop.svg");
     background-size: contain;
   }
 }
