+++ packages/frontend/src/components/chart/chart.scss
@@ -290,7 +290,7 @@
 }
 
 .arrow {
-  background-image: url(../../images/svg/arrow.svg);
+  background-image: url(../../assets/icons/arrow.svg);
   display: block;
   content: '';
   contain: content;
