+++ packages/frontend/src/components/currency/currency.scss
@@ -7,7 +7,7 @@
   border-radius: 1.875rem;
   color: var(--white);
   background-color: var(--accent-blue);
-  background-image: url('../../images/svg/wave-mobile.svg');
+  background-image: url('../../assets/icons/wave-mobile.svg');
   background-size: contain;
   background-repeat: no-repeat;
   background-position: bottom;
@@ -15,12 +15,12 @@
   @media (min-width: 500px) and (max-width: 767px) {
     width: 100%;
     background-size: contain;
-    background-image: url('../../images/svg/wave-tablet.svg');
+    background-image: url('../../assets/icons/wave-tablet.svg');
   }
 
   @media (min-width: $TabletWidth) and (max-width: 1279px) {
     max-width: 336px;
-    background-image: url('../../images/svg/wave-desktop.svg');
+    background-image: url('../../assets/icons/wave-desktop.svg');
     background-size: contain;
   }
 }
