+++ packages/frontend/src/components/table/table.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { selectError } from '../../redux/finance/selectors';
 import { useSelector } from 'react-redux';
 
-import { ReactComponent as EditIcon } from '../../images/svg/edit_icon.svg';
+import { ReactComponent as EditIcon } from '../../assets/icons/edit_icon.svg';
 import useMediaQuery from '@mui/material/useMediaQuery';
 import { CustomTablePagination } from './table.styles.js';
