+++ packages/frontend/src/components/wallet/wallet.jsx
@@ -1,4 +1,4 @@
-import wallet from '../../images/wallet.svg';
+import wallet from '../../assets/icons/wallet.svg';
 
 export const Wallet = ({ width = 30, height = 30 }) => {
   const style = {
